Remove most warnings when compiling

+++ content/geometry/Point.h
@@ -9,7 +9,7 @@
  */
 #pragma once
 
-template <class T> int sgn(T x) { return (x > 0) - (x < 0); }
+template <class T> ll sgn(T x) { return (x > 0) - (x < 0); }
 template<class T>
 struct Point {
 	typedef Point P;
