+++ content/various/Knapsack.h
@@ -9,21 +9,21 @@
  */
 #pragma once
 
-pair<int, vi> knapsack(vi& v, vi& w, int W) {
+pair<ll, vi> knapsack(vi& v, vi& w, ll W) {
     vector<vi> dp(sz(v) + 1, vi(W + 1, 0));
     vector<vi> keep(sz(v) + 1, vi(W + 1, 0));
     rep(i, 1, sz(v) + 1) rep(j, 0, W+1) {
         dp[i][j] = dp[i-1][j];
-        int val = j >= w[i-1] ? dp[i-1][j-w[i-1]] + v[i-1] : 0;
+        ll val = j >= w[i-1] ? dp[i-1][j-w[i-1]] + v[i-1] : 0;
         if (j >= w[i-1] && dp[i][j] < val)
             dp[i][j] = val, keep[i][j] = 1;
     }
-    int V = 0;
-    int j = 0;
+    ll V = 0;
+    ll j = 0;
     rep(i, 0, W+1) if (V < dp[sz(v)][i])
         V = dp[sz(v)][i], j = i;
     vi res;
-    for (int i = sz(v); i > 0; i--) {
+    for (ll i = sz(v); i > 0; i--) {
         if (keep[i][j]) {
             j -= w[i-1];
             res.push_back(i-1);
