+++ stress-tests/geometry/LineIntersection.cpp
@@ -5,7 +5,7 @@
 
 int main() {
 	rep(t,0,1000000) {
-		const int GRID=10;
+		const ll GRID=10;
 		Point<double>
 			a(rand()%GRID, rand()%GRID),
 			b(rand()%GRID, rand()%GRID),
