+++ stress-tests/utilities/template.h
@@ -1,9 +1,9 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-#define rep(i, from, to) for (int i = from; i < (to); ++i)
+#define rep(i, from, to) for (ll i = from; i < (to); ++i)
 #define all(x) x.begin(), x.end()
-#define sz(x) (int)(x).size()
+#define sz(x) (ll)(x).size()
 typedef long long ll;
-typedef pair<int, int> pii;
-typedef vector<int> vi;
+typedef pair<ll, ll> pii;
+typedef vector<ll> vi;
